+++ Plugin.cs
@@ -33,10 +33,14 @@ namespace LethalWashing
         // Configs
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public static ConfigEntry<float> configWashTime;
-        public static ConfigEntry<Vector3> configWorldPosition;
-        public static ConfigEntry<Quaternion> configWorldRotation;
         public static ConfigEntry<int> configMaxItemsInMachine;
         public static ConfigEntry<bool> configEnableDebugging;
+        public static ConfigEntry<bool> configUseDespawnScript;
+
+        public static ConfigEntry<Vector3> configWorldPosition;
+        public static ConfigEntry<Quaternion> configWorldRotation;
+        public static ConfigEntry<Vector3> configGaletryWorldPosition;
+        public static ConfigEntry<Quaternion> configGaletryWorldRotation;
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 
 
@@ -57,13 +61,21 @@ namespace LethalWashing
 
             // General
             configWashTime = Config.Bind("General", "Wash Time", 10f, "Time it takes for the washing machine to finish washing scrap.");
-            configWorldPosition = Config.Bind("General", "World Position", new Vector3(-27.6681f, -2.5747f, -24.764f), "The world spawn position of the washing machine at the company. By default it spawns next to the sell counter.");
-            configWorldRotation = Config.Bind("General", "World Rotation", Quaternion.Euler(0f, 90f, 0f), "The world spawn rotation of the washing machine at the company.");
             configMaxItemsInMachine = Config.Bind("General", "Max items in machine", 5, "How many items can fit in the machine at a time");
             configEnableDebugging = Config.Bind("General", "Enable Debugging", false, "Shows debugging logs in the console");
+            configUseDespawnScript = Config.Bind("General", "Use Despawn Script", false, "In order to prevent items from despawning from the ship on team wipe by default, this mod will check if the item is worth a value of 0 to determine if the item should not despawn. If this is enabled, washing a scrap will attach a script to the item and it will check if the script is on the item to determine if it should not despawn instead. (reloading a save will cause the script to be removed so youll have to rewash items again)");
+
+            configWorldPosition = Config.Bind("General", "World Position", new Vector3(-27.6681f, -2.5747f, -24.764f), "The world spawn position of the washing machine at the company. By default it spawns next to the sell counter.");
+            configWorldRotation = Config.Bind("General", "World Rotation", Quaternion.Euler(0f, 90f, 0f), "The world spawn rotation of the washing machine at the company.");
+
+            configGaletryWorldPosition = Config.Bind("General", "Galetry World Position", new Vector3(-65.2742f, 1.1536f, 20.3886f), "The world spawn position of the washing machine at Galetry. By default it spawns next to the sell counter.");
+            configGaletryWorldRotation = Config.Bind("General", "Galetry World Rotation", Quaternion.Euler(0f, 180f, 0f), "The world spawn rotation of the washing machine at Galetry.");
 
             WashingMachine.worldPosition = configWorldPosition.Value;
             WashingMachine.worldRotation = configWorldRotation.Value;
+
+            WashingMachine.worldPositionGaletry = configGaletryWorldPosition.Value;
+            WashingMachine.worldRotationGaletry = configGaletryWorldRotation.Value;
 
             // Loading Assets
             string sAssemblyLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);